requirements.txt
anthropic==0.46.0
beautifulsoup4==4.13.3
numpy==2.2.3
openai==1.63.2
pandas==2.2.3
Pillow==11.1.0
protobuf==5.29.3
qwen_vl_utils==0.0.10
Requests==2.32.3
rtpt==0.0.4
tqdm==4.66.2
